Move analysis in application.go -> stage.go
