+++ pkg/app/pipedv1/plugin/analysis/config/stage_test.go
@@ -1,0 +1,49 @@
+// Copyright 2025 The PipeCD Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package config
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestValidateAnalysisTemplateRef(t *testing.T) {
+	testcases := []struct {
+		name    string
+		tplName string
+		wantErr bool
+	}{
+		{
+			name:    "valid",
+			tplName: "name",
+			wantErr: false,
+		},
+		{
+			name:    "invalid due to empty template name",
+			tplName: "",
+			wantErr: true,
+		},
+	}
+	for _, tc := range testcases {
+		t.Run(tc.name, func(t *testing.T) {
+			a := &AnalysisTemplateRef{
+				Name: tc.tplName,
+			}
+			err := a.Validate()
+			assert.Equal(t, tc.wantErr, err != nil)
+		})
+	}
+}
