+++ pkg/configv1/application_test.go
@@ -213,34 +213,6 @@ func TestFindSlackUsersAndGroups(t *testing.T) {
 	}
 }
 
-func TestValidateAnalysisTemplateRef(t *testing.T) {
-	testcases := []struct {
-		name    string
-		tplName string
-		wantErr bool
-	}{
-		{
-			name:    "valid",
-			tplName: "name",
-			wantErr: false,
-		},
-		{
-			name:    "invalid due to empty template name",
-			tplName: "",
-			wantErr: true,
-		},
-	}
-	for _, tc := range testcases {
-		t.Run(tc.name, func(t *testing.T) {
-			a := &AnalysisTemplateRef{
-				Name: tc.tplName,
-			}
-			err := a.Validate()
-			assert.Equal(t, tc.wantErr, err != nil)
-		})
-	}
-}
-
 func TestValidateEncryption(t *testing.T) {
 	testcases := []struct {
 		name             string
